fix: feedback page params validation
